requirements.txt
github3.py
GitPython
ipython
